restructured files. Added socketcom class from Dom
